+++ Sources/Scaletor/Protocols/Mode.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+public protocol Mode {
+    var chords: [ChordVoicing] { get }
+    var intervals: [Interval] { get }
+}